requirements.txt
pillow==10.2.0
tk==0.1.0
easy-pil==0.3.2
requests==2.31.0
customtkinter==5.2.1
packaging==23.2
pyinstaller==6.4.0